+++ sdmx/testing/__init__.py
@@ -1,39 +1,32 @@
 import logging
 import os
 from collections import ChainMap
-from contextlib import contextmanager
-from pathlib import Path, PurePosixPath
-from typing import Union
+from pathlib import Path
+from typing import TYPE_CHECKING, Union
 
 import numpy as np
 import pandas as pd
 import pytest
+from xdist import is_xdist_worker
 
 from sdmx.exceptions import HTTPError
 from sdmx.rest import Resource
-from sdmx.source import DataContentType, add_source, sources
+from sdmx.session import Session
+from sdmx.source import DataContentType, Source, sources
 from sdmx.testing.report import ServiceReporter
+from sdmx.util.requests import offline
+
+from . import data
+
+if TYPE_CHECKING:
+    import pytest
 
 log = logging.getLogger(__name__)
 
-
-# Expected to_pandas() results for data files; see expected_data()
-# - Keys are the file name (above) with '.' -> '-': 'foo.xml' -> 'foo-xml'
-# - Data is stored in expected/{KEY}.txt
-# - Values are either argument to pd.read_csv(); or a dict(use='other-key'),
-#   in which case the info for other-key is used instead.
-EXPECTED = {
-    "ng-flat-xml": dict(index_col=[0, 1, 2, 3, 4, 5]),
-    "ng-ts-gf-xml": dict(use="ng-flat-xml"),
-    "ng-ts-xml": dict(use="ng-flat-xml"),
-    "ng-xs-xml": dict(index_col=[0, 1, 2, 3, 4, 5]),
-    # Excluded: this file contains two DataSets, and expected_data() currently
-    # only supports specimens with one DataSet
-    # 'action-delete-json': dict(header=[0, 1, 2, 3, 4]),
-    "xs-json": dict(index_col=[0, 1, 2, 3, 4, 5]),
-    "flat-json": dict(index_col=[0, 1, 2, 3, 4, 5]),
-    "ts-json": dict(use="flat-json"),
-}
+# Pytest stash keys
+KEY_DATA = pytest.StashKey[Path]()
+KEY_SPECIMENS = pytest.StashKey["data.SpecimenCollection"]()
+KEY_SOURCE = pytest.StashKey["Source"]()
 
 
 def assert_pd_equal(left, right, **kwargs):
@@ -49,11 +42,16 @@ def assert_pd_equal(left, right, **kwargs):
 
 
 def pytest_addoption(parser):
-    """Add the ``--sdmx-test-data`` command-line option to pytest."""
+    """Add pytest command-line options."""
+    parser.addoption(
+        "--sdmx-fetch-data",
+        action="store_true",
+        help="fetch test specimens from GitHub",
+    )
     parser.addoption(
         "--sdmx-test-data",
         # Use the environment variable value by default
-        default=os.environ.get("SDMX_TEST_DATA", None),
+        default=os.environ.get("SDMX_TEST_DATA", data.DEFAULT_DIR),
         help="path to SDMX test specimens",
     )
 
@@ -69,29 +67,34 @@ def pytest_configure(config):
     config._sdmx_reporter = ServiceReporter(config)
     config.pluginmanager.register(config._sdmx_reporter)
 
-    # Check the value can be converted to a path, and exists
-    message = "Give --sdmx-test-data=… or set the SDMX_TEST_DATA environment variable"
+    # Convert the option value to an Path instance attribute on `config`
     try:
-        sdmx_test_data = Path(config.option.sdmx_test_data)
+        config.stash[KEY_DATA] = Path(config.option.sdmx_test_data)
     except TypeError:  # pragma: no cover
-        raise RuntimeError(message) from None
-    else:  # pragma: no cover
-        if not sdmx_test_data.exists():  # pragma: no cover
-            # Cannot proceed further; this exception kills the test session
-            raise FileNotFoundError(f"SDMX test data in {sdmx_test_data}\n{message}")
+        raise RuntimeError(data.ERROR) from None
 
-    setattr(config, "sdmx_test_data", sdmx_test_data)
-    setattr(config, "sdmx_specimens", SpecimenCollection(sdmx_test_data))
 
-    # Add a source globally for use with mock_service_adapter
-    # TODO Deduplicate between this and the "testsource" fixture
-    info = dict(
-        id="MOCK",
-        name="Mock source",
-        url="mock://example.com/",
+def pytest_sessionstart(session: "pytest.Session") -> None:
+    """Create session-wide objects.
+
+    These are used by the fixtures :func:`.specimen`, :func:`.testsource`.
+    """
+    c = session.config
+
+    # Create a SpecimenCollection from the files in the directory
+    c.stash[KEY_SPECIMENS] = data.SpecimenCollection(
+        c.stash[KEY_DATA],
+        # Fetch the test data if so configured, and not in an xdist worker process
+        c.option.sdmx_fetch_data and not is_xdist_worker(session),
+    )
+
+    # Create a test source
+    c.stash[KEY_SOURCE] = Source(
+        id="TEST",
+        name="Test source",
+        url="https://example.com/sdmx-rest",
         supports={feature: True for feature in list(Resource)},
     )
-    add_source(info)
 
 
 def pytest_generate_tests(metafunc):
@@ -99,20 +102,10 @@ def pytest_generate_tests(metafunc):
 
     Calls both :func:`parametrize_specimens` and :func:`generate_endpoint_tests`.
     """
-    parametrize_specimens(metafunc)
+    # Handle the 'parametrize_specimens' mark
+    metafunc.config.stash[KEY_SPECIMENS].parametrize(metafunc)
+    # Handle the 'endpoint' fixture request
     generate_endpoint_tests(metafunc)
-
-
-def parametrize_specimens(metafunc):
-    """Handle ``@pytest.mark.parametrize_specimens(…)``."""
-    try:
-        mark = next(metafunc.definition.iter_markers("parametrize_specimens"))
-    except StopIteration:
-        return
-
-    metafunc.parametrize(
-        mark.args[0], metafunc.config.sdmx_specimens.as_params(**mark.kwargs)
-    )
 
 
 #: Marks for use below.
@@ -157,7 +150,11 @@ def generate_endpoint_tests(metafunc):  # noqa: C901  TODO reduce complexity 11 
 
     # Use the test class' source_id attr to look up the Source class
     cls = metafunc.cls
-    source = sources[cls.source_id]
+    source = (
+        sources[cls.source_id]
+        if cls.source_id != "TEST"
+        else metafunc.config.stash[KEY_SOURCE]
+    )
 
     # Merge subclass-specific and "common" xfail marks, preferring the former
     xfails = ChainMap(cls.xfail, cls.xfail_common)
@@ -228,245 +225,58 @@ class MessageTest:
         return sdmx.read_sdmx(path / self.filename)
 
 
-class SpecimenCollection:
-    """Collection of test specimens."""
+@pytest.fixture(scope="session")
+def session_with_pytest_cache(pytestconfig):
+    """Fixture:  A :class:`.Session` that caches within :file:`.pytest_cache`.
 
-    # Path to specimen; file format; data/structure
-    # TODO add version
-    specimens: list[tuple[Path, str, str]]
-
-    def __init__(self, base_path):
-        self.base_path = base_path
-        self.specimens = []
-
-        # XML data files for the ECB exchange rate data flow
-        for source_id in ("ECB_EXR",):
-            for path in base_path.joinpath(source_id).rglob("*.xml"):
-                kind = "data"
-                if "structure" in path.name or "common" in path.name:
-                    kind = "structure"
-                self.specimens.append((path, "xml", kind))
-
-        # JSON data files for ECB and OECD data flows
-        for source_id in ("ECB_EXR", "OECD"):
-            self.specimens.extend(
-                (fp, "json", "data")
-                for fp in base_path.joinpath(source_id).rglob("*.json")
-            )
-
-        # Miscellaneous XML data files
-        self.specimens.extend(
-            (base_path.joinpath(*parts), "xml", "data")
-            for parts in [
-                ("INSEE", "CNA-2010-CONSO-SI-A17.xml"),
-                ("INSEE", "IPI-2010-A21.xml"),
-                ("ESTAT", "esms.xml"),
-                ("ESTAT", "footer.xml"),
-                ("ESTAT", "NAMA_10_GDP-ss.xml"),
-            ]
-        )
-
-        # Miscellaneous XML structure files
-        self.specimens.extend(
-            (base_path.joinpath(*parts), "xml", "structure")
-            for parts in [
-                ("BIS", "actualconstraint-0.xml"),
-                ("BIS", "hierarchicalcodelist-0.xml"),
-                ("ECB", "orgscheme.xml"),
-                ("ECB", "structureset-0.xml"),
-                ("ESTAT", "apro_mk_cola-structure.xml"),
-                ("ESTAT", "esms-structure.xml"),
-                ("ESTAT", "GOV_10Q_GGNFA.xml"),
-                ("ESTAT", "HCL_WSTATUS_SCL_BNSPART.xml"),
-                ("ESTAT", "HCL_WSTATUS_SCL_WSTATUSPR.xml"),
-                ("IAEG-SDGs", "metadatastructure-0.xml"),
-                ("IMF", "1PI-structure.xml"),
-                ("IMF", "CL_AREA-structure.xml"),
-                # Manually reduced subset of the response for this DSD. Test for
-                # <str:CubeRegion> containing both <com:KeyValue> and <com:Attribute>
-                ("IMF", "ECOFIN_DSD-structure.xml"),
-                ("IMF", "hierarchicalcodelist-0.xml"),
-                ("IMF", "structureset-0.xml"),
-                ("IMF_STA", "availableconstraint_CPI.xml"),  # khaeru/sdmx#161
-                ("IMF_STA", "DSD_GFS.xml"),  # khaeru/sdmx#164
-                ("INSEE", "CNA-2010-CONSO-SI-A17-structure.xml"),
-                ("INSEE", "dataflow.xml"),
-                ("INSEE", "gh-205.xml"),
-                ("INSEE", "IPI-2010-A21-structure.xml"),
-                ("ISTAT", "22_289-structure.xml"),
-                ("ISTAT", "47_850-structure.xml"),
-                ("ISTAT", "actualconstraint-0.xml"),
-                ("ISTAT", "metadataflow-0.xml"),
-                ("ISTAT", "metadatastructure-0.xml"),
-                ("OECD", "actualconstraint-0.xml"),
-                ("OECD", "metadatastructure-0.xml"),
-                ("UNICEF", "GLOBAL_DATAFLOW-structure.xml"),
-                ("UNSD", "codelist_partial.xml"),
-                ("SDMX", "HCL_TEST_AREA.xml"),
-                ("SGR", "common-structure.xml"),
-                ("SGR", "hierarchicalcodelist-0.xml"),
-                ("SGR", "metadatastructure-0.xml"),
-                ("SPC", "actualconstraint-0.xml"),
-                ("SPC", "metadatastructure-0.xml"),
-                ("TEST", "gh-142.xml"),
-                ("TEST", "gh-149.xml"),
-                ("WB", "gh-78.xml"),
-            ]
-        )
-
-        # Add files from the SDMX 2.1 specification
-        v21 = base_path.joinpath("v21", "xml")
-        self.specimens.extend((p, "xml", None) for p in v21.glob("**/*.xml"))
-
-        # Add files from the SDMX 3.0 specification
-        v3 = base_path.joinpath("v3")
-
-        # SDMX-CSV
-        self.specimens.extend(
-            (p, "csv", "data") for p in v3.joinpath("csv").glob("*.csv")
-        )
-
-        # commented: SDMX-JSON 2.0 is not yet implemented
-        # # SDMX-JSON
-        # self.specimens.extend(
-        #     (p, "json", "data") for p in v3.joinpath("json", "data").glob("*.json")
-        # )
-        # for dir in ("metadata", "structure"):
-        #     self.specimens.extend(
-        #         (p, "json", "structure")
-        #         for p in v3.joinpath("json", dir).glob("*.json")
-        #     )
-
-        # SDMX-ML
-        self.specimens.extend((p, "xml", None) for p in v3.glob("xml/*.xml"))
-
-    @contextmanager
-    def __call__(self, pattern="", opened=True):
-        """Open the test specimen file with `pattern` in the name."""
-        for path, f, k in self.specimens:
-            if path.match("*" + pattern + "*"):
-                yield open(path, "br") if opened else path
-                return
-        raise ValueError(pattern)  # pragma: no cover
-
-    def as_params(self, format=None, kind=None, marks=dict()):
-        """Generate :func:`pytest.param` from specimens.
-
-        One :func:`~.pytest.param` is generated for each specimen that matches the
-        `format` and `kind` arguments (if any). Marks are attached to each param from
-        `marks`, wherein the keys are partial paths.
-        """
-        # Transform `marks` into a platform-independent mapping from path parts
-        _marks = {PurePosixPath(k).parts: v for k, v in marks.items()}
-
-        for path, f, k in self.specimens:
-            if (format and format != f) or (kind and kind != k):
-                continue
-            p_rel = path.relative_to(self.base_path)
-            yield pytest.param(
-                path,
-                id=str(p_rel),  # String ID for this specimen
-                marks=_marks.get(p_rel.parts, tuple()),  # Look up marks via path parts
-            )
-
-    def expected_data(self, path):
-        """Return the expected :func:`.to_pandas()` result for the specimen `path`."""
-        try:
-            key = path.name.replace(".", "-")
-            info = EXPECTED[key]
-            if "use" in info:
-                # Use the same expected data as another file
-                key = info["use"]
-                info = EXPECTED[key]
-        except KeyError:
-            return None
-
-        args = dict(sep=r"\s+", index_col=[0], header=[0])
-        args.update(info)
-
-        result = pd.read_csv(
-            self.base_path.joinpath("expected", key).with_suffix(".txt"), **args
-        )
-
-        # A series; unwrap
-        if set(result.columns) == {"value"}:
-            result = result["value"]
-
-        return result
+    This subdirectory is ephemeral, and tests **must** pass whether or not it exists and
+    is populated.
+    """
+    p = pytestconfig.cache.mkdir("sdmx-requests-cache")
+    yield Session(cache_name=str(p), backend="filesystem")
 
 
 @pytest.fixture(scope="session")
-def test_data_path(pytestconfig):
-    """Fixture: the :py:class:`.Path` given as --sdmx-test-data."""
-    yield pytestconfig.sdmx_test_data
+def session_with_stored_responses(pytestconfig):
+    """Fixture: A :class:`.Session` returns only stored responses from sdmx-test-data.
+
+    This session (a) uses the 'filesystem' :mod:`requests_cache` backend and (b) is
+    treated with :func:`.offline`, so that *only* stored responses can be returned.
+    """
+    session = Session(backend="memory")
+
+    data.add_responses(
+        session,
+        pytestconfig.stash[KEY_DATA].joinpath("recorded"),
+        pytestconfig.stash[KEY_SOURCE],
+    )
+
+    # Raise an exception on any actual attempts to access the network
+    offline(session)
+
+    yield session
 
 
 @pytest.fixture(scope="session")
 def specimen(pytestconfig):
     """Fixture: the :class:`SpecimenCollection`."""
-    yield pytestconfig.sdmx_specimens
+    yield pytestconfig.stash[KEY_SPECIMENS]
 
 
 @pytest.fixture(scope="session")
-def mock_service_adapter():
-    from requests_mock import Adapter
-
-    import sdmx
-    from sdmx.format import MediaType
-    from sdmx.message import StructureMessage
-
-    a = Adapter()
-
-    common = dict(
-        content=sdmx.to_xml(StructureMessage()),
-        status_code=200,
-        headers={"Content-Type": repr(MediaType("generic", "xml", "2.1"))},
-    )
-    for path in (
-        "actualconstraint/MOCK/all/latest",
-        "agencyscheme/MOCK/all/latest",
-        "allowedconstraint/MOCK/all/latest",
-        "attachementconstraint/MOCK/all/latest",
-        "availableconstraint",
-        "categorisation/MOCK/all/latest",
-        "categoryscheme/MOCK/all/latest",
-        "codelist/MOCK/all/latest",
-        "conceptscheme/MOCK/all/latest",
-        "contentconstraint/MOCK/all/latest",
-        "customtypescheme/MOCK/all/latest",
-        "dataconsumerscheme/MOCK/all/latest",
-        "dataflow/MOCK/all/latest",
-        "dataproviderscheme/MOCK/all/latest",
-        "datastructure/MOCK/all/latest",
-        "hierarchicalcodelist/MOCK/all/latest",
-        "metadataflow/MOCK/all/latest",
-        "metadatastructure/MOCK/all/latest",
-        "namepersonalisationscheme/MOCK/all/latest",
-        "organisationscheme/MOCK/all/latest",
-        "organisationunitscheme/MOCK/all/latest",
-        "process/MOCK/all/latest",
-        "provisionagreement/MOCK/all/latest",
-        "reportingtaxonomy/MOCK/all/latest",
-        "rulesetscheme/MOCK/all/latest",
-        "schema/datastructure/MOCK/all/latest",
-        "structure/MOCK/all/latest",
-        "structureset/MOCK/all/latest",
-        "transformationscheme/MOCK/all/latest",
-        "userdefinedoperatorscheme/MOCK/all/latest",
-        "vtlmappingscheme/MOCK/all/latest",
-    ):
-        a.register_uri("GET", f"mock://example.com/{path}", **common)
-
-    yield a
+def test_data_path(pytestconfig):
+    """Fixture: the :py:class:`.Path` given as --sdmx-test-data."""
+    yield pytestconfig.stash[KEY_DATA]
 
 
 @pytest.fixture(scope="class")
-def testsource():
-    """Fixture: the :attr:`.Source.id` of a non-existent data source."""
-    id = "TEST"
-    add_source(dict(id=id, name="Test source", url="https://example.com/sdmx-rest"))
+def testsource(pytestconfig):
+    """Fixture: the :attr:`.Source.id` of a temporary data source."""
+    s = pytestconfig.stash[KEY_SOURCE]
+
+    sources[s.id] = s
 
     try:
-        yield id
+        yield s.id
     finally:
-        sources.pop(id)
+        sources.pop(s.id)