+++ sdmx/tests/format/test_format_xml.py
@@ -3,6 +3,7 @@ import re
 from pathlib import Path
 
 import pytest
+import responses
 
 import sdmx
 from sdmx.format import Version, xml
@@ -37,17 +38,26 @@ def mock_gh_api():
     For each API endpoint URL queried by :func:.`_gh_zipball`, return a pared-down JSON
     response that contains the required "zipball_url" key.
     """
-    import requests_mock
-
     base = "https://api.github.com/repos/sdmx-twg/sdmx-ml"
 
-    with requests_mock.Mocker(real_http=True) as m:
-        for v in "2.1", "3.0":
-            m.get(
-                f"{base}/releases/tags/v{v}",
-                json=dict(zipball_url=f"{base}/zipball/v{v}"),
-            )
+    # TODO Improve .util.requests to provide (roughly) the same functionality, then drop
+    # use of responses here
+    mock = responses.RequestsMock(assert_all_requests_are_fired=False)
+    mock.add_passthru(re.compile(rf"{base}/zipball/\w+"))
+    mock.add_passthru(re.compile(r"https://codeload.github.com/\w+"))
+
+    for v in "2.1", "3.0", "3.0.0":
+        mock.get(
+            url=f"{base}/releases/tags/v{v}",
+            json=dict(zipball_url=f"{base}/zipball/v{v}"),
+        )
+
+    mock.start()
+
+    try:
         yield
+    finally:
+        mock.stop()
 
 
 @pytest.fixture(scope="module")