+++ sdmx/tests/test_client.py
@@ -5,10 +5,9 @@ from io import BytesIO
 
 import pandas as pd
 import pytest
-import requests_mock
 
 import sdmx
-from sdmx.source import sources
+from sdmx.util.requests import save_response
 
 
 def test_deprecated_request(caplog):
@@ -44,9 +43,9 @@ def test_read_sdmx(tmp_path, specimen):
 
 class TestClient:
     @pytest.fixture
-    def client(self, testsource):
+    def client(self, testsource, session_with_stored_responses):
         """A :class:`Client` connected to a non-existent test source."""
-        return sdmx.Client(testsource)
+        return sdmx.Client(testsource, session=session_with_stored_responses)
 
     def test_init(self):
         with pytest.warns(
@@ -62,7 +61,7 @@ class TestClient:
     def test_clear_cache(self, client):
         client.clear_cache()
 
-    def test_session_attrs(self, caplog, client):
+    def test_session_attrs0(self, caplog, client):
         # Deprecated attributes
         with pytest.warns(DeprecationWarning, match="Setting Client.timeout"):
             client.timeout = 300
@@ -95,6 +94,10 @@ class TestClient:
             "Client.get() args {'allow_redirects': True, 'timeout': 123} replace "
             "{'allow_redirects': False}" in caplog.messages
         )
+
+    def test_session_attrs1(self, testsource, session_with_stored_responses):
+        with pytest.raises(ValueError):
+            sdmx.Client(testsource, session=session_with_stored_responses, verify=False)
 
     def test_dir(self, client):
         """dir() includes convenience methods for resource endpoints."""
@@ -155,24 +158,23 @@ class TestClient:
             client.get("datastructure", validate=False, dry_run=True)
 
     # TODO update or remove
-    @pytest.mark.skip(reason="SDMX 3.0.0 now supported")
-    def test_v3_unsupported(self, testsource, client):
+    @pytest.mark.xfail(reason="SDMX 3.0.0 is now supported → no exception raised")
+    def test_v3_unsupported(self, client):
         """Client raises an exception when an SDMX 3.0 message is returned."""
-        mock = requests_mock.Mocker()
-        df_id = "DATAFLOW"
-        key = ".KEY2.KEY3..KEY5"
-        url = f"{sources[testsource].url}/data/{df_id}/{key}"
-        mock.get(
-            url,
-            body="",
-            headers={"Content-Type": "application/vnd.sdmx.data+xml; version=3.0.0"},
+        df_id, key = "DATAFLOW", ".KEY2.KEY3..KEY5"
+
+        save_response(
+            client.session,
+            "GET",
+            url=f"{client.source.url}/data/{df_id}/{key}",
+            content="""<?xml version='1.0' encoding='UTF-8'?>
+<mes:Structure xmlns:mes="http://www.sdmx.org/resources/sdmxml/schemas/v3_0/message">
+</mes:Structure>""".encode(),
+            headers={"Content-type": "application/vnd.sdmx.data+xml; version=3.0.0"},
         )
 
-        with (
-            mock,
-            pytest.raises(
-                ValueError, match="can't determine a reader for response content type"
-            ),
+        with pytest.raises(
+            ValueError, match="can't determine a reader for response content type"
         ):
             client.get("data", resource_id=df_id, key=key)
 
@@ -206,7 +208,7 @@ def test_request_get_args():
             "UNE_RT_A",
             key={"geo": "EL+ES+IE"},
             params={"startPeriod": "2007"},
-            provider="ESTAT",
+            agency_id="ESTAT",
         )
 
     # Using an unknown endpoint is an exception